% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dumps_to_file.R
\name{extract_wikis}
\alias{extract_wikis}
\title{function that extracts wikis from files}
\usage{
extract_wikis(pagecount_txt, directory = NULL)
}
\description{
function that extracts wikis from files
}
