% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dumps_to_file.R
\name{filter_dumps_to_file_par}
\alias{filter_dumps_to_file_par}
\title{filter dumps to file}
\usage{
filter_dumps_to_file_par(flist, wiki, directory = NULL)
}
\arguments{
\item{wiki}{wiki to filter for}

\item{directory}{base dir where to store filtering}

\item{fname}{file to filter through}
}
\description{
filter dumps to file
}
