% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dumps_to_file.R
\name{filter_worker}
\alias{filter_worker}
\title{worker function}
\usage{
filter_worker(fname, wiki, directory = NULL)
}
\arguments{
\item{wiki}{wiki to filter for}

\item{directory}{base dir where to store filtering}

\item{flist}{list of files to filter through}
}
\description{
worker function
}
