% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dumps.R
\name{get_dumps}
\alias{get_dumps}
\title{function that downloads dump files for specific dates}
\usage{
get_dumps(ts, directory = NULL)
}
\arguments{
\item{ts}{timestamp e.g. "20071201" or shorter that will be used to match dump links}

\item{directory}{directory to store downloads in}
}
\description{
function that downloads dump files for specific dates
}
