% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dump_links.R
\name{get_dump_links}
\alias{get_dump_links}
\title{get links for dump downloads}
\usage{
get_dump_links(force = FALSE, directory = NULL)
}
\arguments{
\item{force}{force re-downloading and re-extracting dump links}

\item{directory}{path to store dump_links.txt file}
}
\description{
get links for dump downloads
}
